random skeleton location

+++ sandbox/haoran/myscripts/modify_variant_json.py
@@ -7,7 +7,8 @@ from rllab.misc.console import colorize
 parser = argparse.ArgumentParser()
 parser.add_argument('prefix', type=str, default='??????',nargs='?')
 parser.add_argument('--param', type=str, default='', nargs='?')
-parser.add_argument('--value', type=str, default='', nargs='?')
+parser.add_argument('--old', type=str, default='', nargs='?')
+parser.add_argument('--new', type=str, default='', nargs='?')
 parser.add_argument('--yes', default=False, action='store_true')
 args = parser.parse_args()
 
@@ -23,22 +24,28 @@ for subdirname in os.listdir(dirname):
 
 for json_file in json_files:
     print("File: %s"%(json_file))
+    confirm = False
     with open(json_file,"r") as f:
         params = json.load(f)
         if args.param in params.keys():
             old_value = params[args.param]
+            value_type = type(old_value)
+            if args.old == '' or value_type(args.old) != old_value:
+                continue
             if args.yes:
                 confirm = True
             else:
+                new_value = value_type(args.new)
                 confirm = query_yes_no("Replace {name}={old_value} by {new_value}?".format(
                     name=args.param,
                     old_value=old_value,
-                    new_value=args.value,
+                    new_value=new_value,
                 ))
             if confirm:
-                params[args.param] = args.value
-    with open(json_file,"w") as f:
-        json.dump(params,f)
-        print(colorize("Changes made and saved.","yellow"))
+                params[args.param] = new_value 
+    if confirm:
+        with open(json_file,"w") as f:
+            json.dump(params,f)
+            print(colorize("Changes made and saved.","yellow"))
